requirements.txt
# Django Framework
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0

# JWT Authentication
djangorestframework-simplejwt>=5.3,<6.0

# Database
psycopg2-binary>=2.9,<3.0

# Cache
redis>=5.0,<6.0
django-redis>=5.4,<6.0

# Task Queue
celery>=5.3,<6.0

# Configuration
python-dotenv>=1.0,<2.0
PyYAML>=6.0,<7.0

# Utilities
python-dateutil>=2.8,<3.0
pytz>=2024.1,<2025.0

# Testing
pytest>=7.4,<8.0
pytest-django>=4.5,<5.0
pytest-cov>=4.1,<5.0
factory-boy>=3.3,<4.0

# Development
ipython>=8.0,<9.0
